# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BusStatus": {
				"enum": [
					"IDLE",
					"MOVING",
					"STOPPED",
					"DELAYED",
					"OFFLINE"
				],
				"type": "string"
			},
			"LocationUpdateDTO": {
				"properties": {
					"busId": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"speed": {
						"type": "number",
						"format": "double"
					},
					"heading": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/BusStatus"
					}
				},
				"required": [
					"busId",
					"latitude",
					"longitude"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateStudentDTO": {
				"properties": {
					"collegeId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"rollNumber": {
						"type": "string"
					},
					"className": {
						"type": "string"
					},
					"section": {
						"type": "string"
					},
					"contact": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"collegeId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Status": {
				"enum": [
					"ACTIVE",
					"INACTIVE",
					"DEACTIVATED"
				],
				"type": "string"
			},
			"UpdateStudentDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"rollNumber": {
						"type": "string"
					},
					"className": {
						"type": "string"
					},
					"section": {
						"type": "string"
					},
					"contact": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AssignStudentBusDTO": {
				"properties": {
					"studentId": {
						"type": "string"
					},
					"busId": {
						"type": "string"
					},
					"collegeId": {
						"type": "string"
					}
				},
				"required": [
					"studentId",
					"busId",
					"collegeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssignmentType": {
				"enum": [
					"PICKUP",
					"DROP",
					"BOTH"
				],
				"type": "string"
			},
			"AssignStudentStopDTO": {
				"properties": {
					"studentId": {
						"type": "string"
					},
					"stopId": {
						"type": "string"
					},
					"collegeId": {
						"type": "string"
					},
					"assignmentType": {
						"$ref": "#/components/schemas/AssignmentType"
					}
				},
				"required": [
					"studentId",
					"stopId",
					"collegeId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateBusDTO": {
				"properties": {
					"busNumber": {
						"type": "string"
					},
					"vehicleNumber": {
						"type": "string"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"driverId": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/BusStatus"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "access_token",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "busapp-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/v1/tracking/update-location": {
			"post": {
				"operationId": "UpdateLocation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update real-time GPS location coordinates of a bus.\n\n### Intent & Business Purpose\nReceives real-time GPS coordinates (latitude, longitude, speed, heading) from driver mobile device or onboard IoT tracking device. Updates live tracking state in Firebase Realtime Database.\n\n### Target Audience & Roles\n- **Allowed Roles:** `DRIVER`.\n\n### Request Body\n- `busId` *(required string)*: Bus vehicle UUID.\n- `latitude` *(required number)*: GPS latitude coordinate.\n- `longitude` *(required number)*: GPS longitude coordinate.\n- `speed` *(optional number)*: Current vehicle speed (km/h).\n- `heading` *(optional number)*: Heading angle / direction in degrees.\n\n### Side Effects\n- Pushes updated coordinates to Firebase Realtime Database node (`buses/{busId}`).",
				"tags": [
					"Live Bus Tracking"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Real-time GPS location update payload.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LocationUpdateDTO",
								"description": "Real-time GPS location update payload."
							}
						}
					}
				}
			}
		},
		"/api/v1/tracking/bus/{busId}": {
			"get": {
				"operationId": "GetBusTracking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get current live GPS location & shift state of a specific bus.\n\n### Intent & Business Purpose\nFetches real-time location, speed, direction, and active shift details for a given bus. Used by students/parents on map views.\n\n### Path Parameters\n- `busId` *(required string)*: Bus UUID.",
				"tags": [
					"Live Bus Tracking"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Bus UUID.",
						"in": "path",
						"name": "busId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/tracking/fleet": {
			"get": {
				"operationId": "GetCollegeFleetTracking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get real-time tracking overview for all buses in a college fleet.\n\n### Intent & Business Purpose\nReturns live location snapshot for all active buses belonging to the college campus. Used by college admins on command center map.\n\n### Target Audience & Roles\n- **Allowed Roles:** `ADMIN`, `COLLEGE_ADMIN`.",
				"tags": [
					"Live Bus Tracking"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Optional college UUID filter.",
						"in": "query",
						"name": "collegeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/students": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "List all student records for a college.\n\n### Intent & Business Purpose\nRetrieves student profiles for a college tenant, including roll number, batch, assigned bus ID, and pickup stop ID.\n\n### Target Audience & Roles\n- **Allowed Roles:** `ADMIN`, `COLLEGE_ADMIN`.",
				"tags": [
					"Students"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Optional college UUID filter for Super Admins.",
						"in": "query",
						"name": "collegeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Register a student profile.\n\n### Request Body\n- `userId` *(required string)*: User account UUID.\n- `rollNo` *(required string)*: Roll number or student ID.\n- `collegeId` *(required string)*: College UUID.\n- `batch` *(optional string)*: Academic batch or year.\n- `busId` *(optional string)*: Bus UUID.\n- `stopId` *(optional string)*: Pickup stop UUID.",
				"tags": [
					"Students"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Student registration payload.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStudentDTO",
								"description": "Student registration payload."
							}
						}
					}
				}
			}
		},
		"/api/v1/students/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get student details by ID.\n\n### Path Parameters\n- `id` *(required string)*: Student UUID.",
				"tags": [
					"Students"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Student UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update student profile.\n\n### Path Parameters\n- `id` *(required string)*: Student UUID.",
				"tags": [
					"Students"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Student UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Update fields.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStudentDTO",
								"description": "Update fields."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Deactivate a student profile.\n\n### Path Parameters\n- `id` *(required string)*: Student UUID.",
				"tags": [
					"Students"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Student UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/students/assign-bus": {
			"post": {
				"operationId": "AssignBus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Assign bus to student.\n\n### Request Body\n- `studentId` *(required string)*: Student UUID.\n- `busId` *(required string)*: Bus UUID.",
				"tags": [
					"Students"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Bus assignment payload.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignStudentBusDTO",
								"description": "Bus assignment payload."
							}
						}
					}
				}
			}
		},
		"/api/v1/students/assign-stop": {
			"post": {
				"operationId": "AssignStop",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Assign pickup stop to student.\n\n### Request Body\n- `studentId` *(required string)*: Student UUID.\n- `stopId` *(required string)*: Route stop UUID.",
				"tags": [
					"Students"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Pickup stop assignment payload.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignStudentStopDTO",
								"description": "Pickup stop assignment payload."
							}
						}
					}
				}
			}
		},
...
		"/api/v1/buses/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get bus details by ID.\n\n### Intent & Business Purpose\nRetrieves complete profile, current status, assigned driver, and assigned route details for a specific bus.\n\n### Target Audience & Roles\n- **Allowed Roles:** Authenticated users belonging to the tenant college.\n\n### Path Parameters\n- `id` *(required string)*: Unique UUID of the bus.\n\n### Edge Cases & QA Testing Focus\n- Accessing a bus ID belonging to a different college tenant returns HTTP 404 Not Found.",
				"tags": [
					"Buses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Unique bus UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update bus details.\n\n### Intent & Business Purpose\nUpdates vehicle information, status (`ACTIVE`, `MAINTENANCE`, `INACTIVE`), or capacity.\n\n### Target Audience & Roles\n- **Allowed Roles:** `ADMIN`, `COLLEGE_ADMIN`.\n\n### Path Parameters\n- `id` *(required string)*: Bus UUID.\n\n### Side Effects\n- Modifies existing `buses` table row.",
				"tags": [
					"Buses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Bus UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Fields to update.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBusDTO",
								"description": "Fields to update."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Soft-delete / deactivate a bus.\n\n### Intent & Business Purpose\nSoft-deactivates a bus record from active operational fleet views.\n\n### Target Audience & Roles\n- **Allowed Roles:** `ADMIN`, `COLLEGE_ADMIN`.\n\n### Path Parameters\n- `id` *(required string)*: Bus UUID.\n\n### Side Effects\n- Sets bus status to `INACTIVE` / soft deletes record.",
				"tags": [
					"Buses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Bus UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:5000",
			"description": "Development server"
		}
	]
}